requirements.txt
asgiref==3.8.1
asttokens==2.4.1
black==24.4.2
cfgv==3.4.0
click==8.1.7
decorator==5.1.1
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.15.4
graphviz==0.20.3
identify==2.6.0
ipython==8.26.0
jedi==0.19.1
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.2
pre-commit==3.7.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.9.0
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.3
typing_extensions==4.12.2
virtualenv==20.26.3
wcwidth==0.2.13
